Fix the validation issue on decidim budget when vote per project is activated.
This is a quick fix and the extend need to be restored in the long term.
