+++ decidim-budgets/app/models/decidim/budgets/order.rb
@@ -16,12 +16,50 @@ module Decidim
 
       validates :user, uniqueness: { scope: :component }
       validate :user_belongs_to_organization
-      validates :total_budget, numericality: { greater_than_or_equal_to: :minimum_budget }, if: :checked_out?
+      validates :total_budget, numericality: { greater_than_or_equal_to: :minimum_budget }, if: :budget_check_out
       validates :total_budget, numericality: { less_than_or_equal_to: :maximum_budget }, unless: :per_project
       validates :total_projects, numericality: { less_than_or_equal_to: :number_of_projects }, if: :per_project
 
       scope :finished, -> { where.not(checked_out_at: nil) }
       scope :pending, -> { where(checked_out_at: nil) }
+
+      def budget_check_out
+        checked_out? && !per_project
+      end
+
+      def per_project
+        component.settings.vote_per_project?
+      end
+
+      def limit_project_reached?
+        return false unless per_project
+        total_projects == number_of_projects
+      end
+
+      def total_projects
+        projects.count
+      end
+
+      def remaining_projects
+        number_of_projects - projects.count
+      end
+
+      def can_checkout?
+        if component.settings.vote_per_project?
+          limit_project_reached?
+        else
+          total_budget.to_f >= minimum_budget
+        end
+      end
+
+      def number_of_projects
+        component.settings.total_projects
+      end
+
+      def maximum_budget
+        return 0 unless component || !per_project
+        component.settings.total_budget.to_f
+      end
 
       # Public: Returns the sum of project budgets
       def total_budget
@@ -33,11 +71,6 @@ module Decidim
         checked_out_at.present?
       end
 
-      # Public: Check if the order total budget is enough to checkout
-      def can_checkout?
-        total_budget.to_f >= minimum_budget
-      end
-
       # Public: Returns the order budget percent from the settings total budget
       def budget_percent
         (total_budget.to_f / component.settings.total_budget.to_f) * 100
@@ -47,12 +80,6 @@ module Decidim
       def minimum_budget
         return 0 unless component
         component.settings.total_budget.to_f * (component.settings.vote_threshold_percent.to_f / 100)
-      end
-
-      # Public: Returns the required maximum budget to checkout
-      def maximum_budget
-        return 0 unless component
-        component.settings.total_budget.to_f
       end
 
       private
