+++ decidim-budgets/config/initializers/extends.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require "decidim/extends/models/project_extend.rb"
-require "decidim/extends/models/order_extend.rb"
 require "decidim/extends/commands/create_project_extend.rb"
 require "decidim/extends/commands/update_project_extend.rb"
 require "decidim/extends/forms/project_form_extend.rb"
